% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az-resource-group.R
\name{get_workspace.az_resource_group}
\alias{get_workspace.az_resource_group}
\title{Get Workspace from Resource Group}
\usage{
\method{get_workspace}{az_resource_group}(x, name, ...)
}
\arguments{
\item{x}{Resource group.}

\item{name}{Name of workspace to access.}

\item{...}{
  Arguments passed on to \code{\link[azuremlsdk:get_workspace]{azuremlsdk::get_workspace}}
  \describe{
    \item{\code{auth}}{The \code{ServicePrincipalAuthentication} or \code{InteractiveLoginAuthentication}
object. For more details refer to https://aka.ms/aml-notebook-auth. If NULL,
the default Azure CLI credentials will be used or the API will prompt for credentials.}
    \item{\code{subscription_id}}{A string of the subscription ID to use. The parameter
is required if the user has access to more than one subscription.}
    \item{\code{resource_group}}{A string of the resource group to use. If \code{NULL} the
method will search all resource groups in the subscription.}
  }}
}
\description{
Uses a resource group's subscription identifier and resource group name to
access an Azure Machine-Learning workspace by workspace name.
}
