% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az-resource-group.R
\name{get_workspace}
\alias{get_workspace}
\title{Get Workspace}
\usage{
get_workspace(x, ...)
}
\arguments{
\item{x}{Something to get a workspace from, typically a resource group.}

\item{...}{Extra parameters to pass.}
}
\description{
Get Workspace
}
