% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{list_resources_by_type}
\alias{list_resources_by_type}
\title{List Resources by Type}
\usage{
list_resources_by_type(x, resource_type, ...)
}
\arguments{
\item{x}{Resource group.}

\item{resource_type}{Type of resource to list.}

\item{...}{Extra parameters passed to \code{list_resources} method of Azure
resource group.}
}
\description{
List Resources by Type
}
