% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az-resource-group.R
\name{list_workspaces.az_resource_group}
\alias{list_workspaces.az_resource_group}
\title{List Machine Learning Workspaces}
\usage{
\method{list_workspaces}{az_resource_group}(x, ...)
}
\arguments{
\item{x}{Resource group.}

\item{...}{Ignored parameters.}
}
\description{
List Machine Learning Workspaces
}
