% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az-resource-group.R
\name{list_workspaces}
\alias{list_workspaces}
\title{List Machine Learning Workspaces}
\usage{
list_workspaces(x, ...)
}
\arguments{
\item{x}{Something to get a workspace from, typically a resource group.}

\item{...}{Extra parameters to pass.}
}
\description{
List Machine Learning Workspaces
}
