% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{register_datastore}
\alias{register_datastore}
\title{Register Data Store}
\usage{
register_datastore(x, ...)
}
\arguments{
\item{x}{Existing Azure blob or file store.}

\item{...}{Extra parameters for registration.}

\item{name}{Name of new store.}
}
\description{
Finds or creates a new store.
}
\details{
Passes the subscription and resource group in order to allow for
\code{grant_workspace_access} as \code{TRUE}. Attempting to grant workspace
access otherwise raises a Python exception response.
}
\examples{
\dontrun{
# Register a new store using the workspace default store.
azml::register_datastore(workspace$get_default_datastore(), "xyz")
}
}
