% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az-resource-group.R
\name{get_workspace.az_resource_group}
\alias{get_workspace.az_resource_group}
\title{Get Workspace from Resource Group}
\usage{
\method{get_workspace}{az_resource_group}(x, name, ...)
}
\arguments{
\item{x}{Resource group.}

\item{name}{Name of workspace to access.}

\item{...}{Ignored parameters.}
}
\description{
Uses a resource group's subscription identifier and resource group name to
access an Azure Machine-Learning workspace by workspace name.
}
